[SEIM][Detection Engine] Moves the io-ts schemas to the common folder from the server side
## Summary

This moves the io-ts schemas from the common folder from the server side up to the common folder.

### Checklist

- [x] [Unit or functional tests](https://github.com/elastic/kibana/blob/master/CONTRIBUTING.md#cross-browser-compatibility) were updated or added to match the most common scenarios
