+++ x-pack/plugins/siem/common/detection_engine/schemas/common/schemas.ts
