+++ x-pack/plugins/siem/common/detection_engine/schemas/response/prepackaged_rules_schema.ts
@@ -7,21 +7,14 @@
 import * as t from 'io-ts';
 
 /* eslint-disable @typescript-eslint/camelcase */
-import {
-  rules_installed,
-  rules_custom_installed,
-  rules_not_installed,
-  rules_not_updated,
-} from './schemas';
+import { rules_installed, rules_updated } from '../common/schemas';
 /* eslint-enable @typescript-eslint/camelcase */
 
-export const prePackagedRulesStatusSchema = t.exact(
+export const prePackagedRulesSchema = t.exact(
   t.type({
-    rules_custom_installed,
     rules_installed,
-    rules_not_installed,
-    rules_not_updated,
+    rules_updated,
   })
 );
 
-export type PrePackagedRulesStatusSchema = t.TypeOf<typeof prePackagedRulesStatusSchema>;
+export type PrePackagedRulesSchema = t.TypeOf<typeof prePackagedRulesSchema>;
