+++ x-pack/plugins/siem/common/detection_engine/schemas/response/rules_bulk_schema.ts
