+++ x-pack/plugins/siem/common/detection_engine/schemas/response/rules_schema.ts
@@ -7,10 +7,11 @@
 /* eslint-disable @typescript-eslint/camelcase */
 import * as t from 'io-ts';
 import { isObject } from 'lodash/fp';
-import { Either, fold, right, left } from 'fp-ts/lib/Either';
-
+import { Either, left, fold } from 'fp-ts/lib/Either';
 import { pipe } from 'fp-ts/lib/pipeable';
-import { checkTypeDependents } from './check_type_dependents';
+import { typeAndTimelineOnlySchema, TypeAndTimelineOnly } from './type_timeline_only_schema';
+import { isMlRule } from '../../../machine_learning/helpers';
+
 import {
   actions,
   anomaly_threshold,
@@ -54,9 +55,8 @@ import {
   filters,
   meta,
   note,
-} from './schemas';
+} from '../common/schemas';
 import { ListsDefaultArray } from '../types/lists_default_array';
-import { hasListsFeature } from '../../../feature_flags';
 
 /**
  * This is the required fields for the rules schema response. Put all required properties on
@@ -155,32 +155,92 @@ export const rulesSchema = new t.Type<
   'RulesSchema',
   (input: unknown): input is RulesWithoutTypeDependentsSchema => isObject(input),
   (input): Either<t.Errors, RulesWithoutTypeDependentsSchema> => {
-    const output = checkTypeDependents(input);
-    if (!hasListsFeature()) {
-      // TODO: (LIST-FEATURE) Remove this after the lists feature is an accepted feature for a particular release
-      return removeList(output);
-    } else {
-      return output;
-    }
+    return checkTypeDependents(input);
   },
   t.identity
 );
-
-// TODO: (LIST-FEATURE) Remove this after the lists feature is an accepted feature for a particular release
-export const removeList = (
-  decoded: Either<t.Errors, RequiredRulesSchema>
-): Either<t.Errors, RequiredRulesSchema> => {
-  const onLeft = (errors: t.Errors): Either<t.Errors, RequiredRulesSchema> => left(errors);
-  const onRight = (decodedValue: RequiredRulesSchema): Either<t.Errors, RequiredRulesSchema> => {
-    delete decodedValue.exceptions_list;
-    return right(decodedValue);
-  };
-  const folded = fold(onLeft, onRight);
-  return pipe(decoded, folded);
-};
 
 /**
  * This is the correct type you want to use for Rules that are outputted from the
  * REST interface. This has all base and all optional properties merged together.
  */
 export type RulesSchema = t.TypeOf<typeof rulesSchema>;
+
+export const addSavedId = (typeAndTimelineOnly: TypeAndTimelineOnly): t.Mixed[] => {
+  if (typeAndTimelineOnly.type === 'saved_query') {
+    return [t.exact(t.type({ saved_id: dependentRulesSchema.props.saved_id }))];
+  } else {
+    return [];
+  }
+};
+
+export const addTimelineTitle = (typeAndTimelineOnly: TypeAndTimelineOnly): t.Mixed[] => {
+  if (typeAndTimelineOnly.timeline_id != null) {
+    return [
+      t.exact(t.type({ timeline_title: dependentRulesSchema.props.timeline_title })),
+      t.exact(t.type({ timeline_id: dependentRulesSchema.props.timeline_id })),
+    ];
+  } else {
+    return [];
+  }
+};
+
+export const addQueryFields = (typeAndTimelineOnly: TypeAndTimelineOnly): t.Mixed[] => {
+  if (typeAndTimelineOnly.type === 'query' || typeAndTimelineOnly.type === 'saved_query') {
+    return [
+      t.exact(t.type({ query: dependentRulesSchema.props.query })),
+      t.exact(t.type({ language: dependentRulesSchema.props.language })),
+    ];
+  } else {
+    return [];
+  }
+};
+
+export const addMlFields = (typeAndTimelineOnly: TypeAndTimelineOnly): t.Mixed[] => {
+  if (isMlRule(typeAndTimelineOnly.type)) {
+    return [
+      t.exact(t.type({ anomaly_threshold: dependentRulesSchema.props.anomaly_threshold })),
+      t.exact(
+        t.type({ machine_learning_job_id: dependentRulesSchema.props.machine_learning_job_id })
+      ),
+    ];
+  } else {
+    return [];
+  }
+};
+
+export const getDependents = (typeAndTimelineOnly: TypeAndTimelineOnly): t.Mixed => {
+  const dependents: t.Mixed[] = [
+    t.exact(requiredRulesSchema),
+    t.exact(partialRulesSchema),
+    ...addSavedId(typeAndTimelineOnly),
+    ...addTimelineTitle(typeAndTimelineOnly),
+    ...addQueryFields(typeAndTimelineOnly),
+    ...addMlFields(typeAndTimelineOnly),
+  ];
+
+  if (dependents.length > 1) {
+    // This unsafe cast is because t.intersection does not use an array but rather a set of
+    // tuples and really does not look like they expected us to ever dynamically build up
+    // intersections, but here we are doing that. Looking at their code, although they limit
+    // the array elements to 5, it looks like you have N number of intersections
+    const unsafeCast: [t.Mixed, t.Mixed] = dependents as [t.Mixed, t.Mixed];
+    return t.intersection(unsafeCast);
+  } else {
+    // We are not allowed to call t.intersection with a single value so we return without
+    // it here normally.
+    return dependents[0];
+  }
+};
+
+export const checkTypeDependents = (input: unknown): Either<t.Errors, RequiredRulesSchema> => {
+  const typeOnlyDecoded = typeAndTimelineOnlySchema.decode(input);
+  const onLeft = (errors: t.Errors): Either<t.Errors, RequiredRulesSchema> => left(errors);
+  const onRight = (
+    typeAndTimelineOnly: TypeAndTimelineOnly
+  ): Either<t.Errors, RequiredRulesSchema> => {
+    const intersections = getDependents(typeAndTimelineOnly);
+    return intersections.decode(input);
+  };
+  return pipe(typeOnlyDecoded, fold(onLeft, onRight));
+};
