+++ x-pack/plugins/siem/common/detection_engine/schemas/types/positive_integer.ts
