+++ x-pack/plugins/siem/common/detection_engine/schemas/types/positive_integer_greater_than_zero.test.ts
@@ -4,33 +4,33 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { RiskScore } from './risk_score';
+import { PositiveIntegerGreaterThanZero } from './positive_integer_greater_than_zero';
 import { pipe } from 'fp-ts/lib/pipeable';
 import { left } from 'fp-ts/lib/Either';
-import { foldLeftRight, getPaths } from '../../../../../../common/test_utils';
+import { foldLeftRight, getPaths } from '../../../test_utils';
 
-describe('risk_score', () => {
+describe('positive_integer_greater_than_zero', () => {
   test('it should validate a positive number', () => {
     const payload = 1;
-    const decoded = RiskScore.decode(payload);
+    const decoded = PositiveIntegerGreaterThanZero.decode(payload);
     const message = pipe(decoded, foldLeftRight);
 
     expect(getPaths(left(message.errors))).toEqual([]);
     expect(message.schema).toEqual(payload);
   });
 
-  test('it should validate a zero', () => {
+  test('it should NOT validate a zero', () => {
     const payload = 0;
-    const decoded = RiskScore.decode(payload);
+    const decoded = PositiveIntegerGreaterThanZero.decode(payload);
     const message = pipe(decoded, foldLeftRight);
 
-    expect(getPaths(left(message.errors))).toEqual([]);
-    expect(message.schema).toEqual(payload);
+    expect(getPaths(left(message.errors))).toEqual(['Invalid value "0" supplied to ""']);
+    expect(message.schema).toEqual({});
   });
 
   test('it should NOT validate a negative number', () => {
     const payload = -1;
-    const decoded = RiskScore.decode(payload);
+    const decoded = PositiveIntegerGreaterThanZero.decode(payload);
     const message = pipe(decoded, foldLeftRight);
 
     expect(getPaths(left(message.errors))).toEqual(['Invalid value "-1" supplied to ""']);
@@ -39,19 +39,10 @@ describe('risk_score', () => {
 
   test('it should NOT validate a string', () => {
     const payload = 'some string';
-    const decoded = RiskScore.decode(payload);
+    const decoded = PositiveIntegerGreaterThanZero.decode(payload);
     const message = pipe(decoded, foldLeftRight);
 
     expect(getPaths(left(message.errors))).toEqual(['Invalid value "some string" supplied to ""']);
-    expect(message.schema).toEqual({});
-  });
-
-  test('it should NOT validate a risk score greater than 100', () => {
-    const payload = 101;
-    const decoded = RiskScore.decode(payload);
-    const message = pipe(decoded, foldLeftRight);
-
-    expect(getPaths(left(message.errors))).toEqual(['Invalid value "101" supplied to ""']);
     expect(message.schema).toEqual({});
   });
 });
