+++ x-pack/plugins/siem/common/detection_engine/schemas/types/postive_integer.test.ts
@@ -4,33 +4,33 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { PositiveIntegerGreaterThanZero } from './positive_integer_greater_than_zero';
+import { PositiveInteger } from './positive_integer';
 import { pipe } from 'fp-ts/lib/pipeable';
 import { left } from 'fp-ts/lib/Either';
-import { getPaths, foldLeftRight } from '../../../../../../common/test_utils';
+import { foldLeftRight, getPaths } from '../../../test_utils';
 
 describe('positive_integer_greater_than_zero', () => {
   test('it should validate a positive number', () => {
     const payload = 1;
-    const decoded = PositiveIntegerGreaterThanZero.decode(payload);
+    const decoded = PositiveInteger.decode(payload);
     const message = pipe(decoded, foldLeftRight);
 
     expect(getPaths(left(message.errors))).toEqual([]);
     expect(message.schema).toEqual(payload);
   });
 
-  test('it should NOT validate a zero', () => {
+  test('it should validate a zero', () => {
     const payload = 0;
-    const decoded = PositiveIntegerGreaterThanZero.decode(payload);
+    const decoded = PositiveInteger.decode(payload);
     const message = pipe(decoded, foldLeftRight);
 
-    expect(getPaths(left(message.errors))).toEqual(['Invalid value "0" supplied to ""']);
-    expect(message.schema).toEqual({});
+    expect(getPaths(left(message.errors))).toEqual([]);
+    expect(message.schema).toEqual(payload);
   });
 
   test('it should NOT validate a negative number', () => {
     const payload = -1;
-    const decoded = PositiveIntegerGreaterThanZero.decode(payload);
+    const decoded = PositiveInteger.decode(payload);
     const message = pipe(decoded, foldLeftRight);
 
     expect(getPaths(left(message.errors))).toEqual(['Invalid value "-1" supplied to ""']);
@@ -39,7 +39,7 @@ describe('positive_integer_greater_than_zero', () => {
 
   test('it should NOT validate a string', () => {
     const payload = 'some string';
-    const decoded = PositiveIntegerGreaterThanZero.decode(payload);
+    const decoded = PositiveInteger.decode(payload);
     const message = pipe(decoded, foldLeftRight);
 
     expect(getPaths(left(message.errors))).toEqual(['Invalid value "some string" supplied to ""']);
