+++ x-pack/plugins/siem/common/detection_engine/schemas/types/risk_score.ts
