+++ x-pack/plugins/siem/common/detection_engine/schemas/types/uuid.ts
