+++ x-pack/plugins/siem/server/lib/detection_engine/routes/rules/create_rules_bulk_route.ts
@@ -6,6 +6,7 @@
 
 import uuid from 'uuid';
 
+import { rulesBulkSchema } from '../../../../../common/detection_engine/schemas/response/rules_bulk_schema';
 import { IRouter } from '../../../../../../../../src/core/server';
 import { DETECTION_ENGINE_RULES_URL } from '../../../../../common/constants';
 import { SetupPlugins } from '../../../../plugin';
@@ -24,7 +25,6 @@ import {
   buildSiemResponse,
 } from '../utils';
 import { createRulesBulkSchema } from '../schemas/create_rules_bulk_schema';
-import { rulesBulkSchema } from '../schemas/response/rules_bulk_schema';
 import { updateRulesNotifications } from '../../rules/update_rules_notifications';
 
 export const createRulesBulkRoute = (router: IRouter, ml: SetupPlugins['ml']) => {
